NxNexusPageHeader component & example page RSC-251
