--- gallery/src/components/NxNexusPageHeader/ComplexPageHeaderExample.tsx
@@ -1,34 +1,0 @@
-/*
- * Copyright (c) 2019-present Sonatype, Inc.
- * This program and the accompanying materials are made available under
- * the terms of the Eclipse Public License 2.0 which accompanies this
- * distribution and is available at https://www.eclipse.org/legal/epl-2.0/.
- */
-import React from 'react';
-
-import { NxButton, NxNexusPageHeader } from '@sonatype/react-shared-components';
-
-const ComplexPageHeaderExample = () => {
-  const links = [{
-    name: 'NxButton',
-    href: '#/pages/NxButton'
-  }, {
-    name: 'NxPageHeader',
-    href: '#/pages/NxPageHeader',
-    current: true
-  }];
-
-  function onButtonClick() {
-    alert('Click');
-  }
-
-  return (
-    <NxNexusPageHeader productInfo={{ name: 'Test Product', version: '1.0.0' }}
-                       homeLink="#"
-                       links={links}>
-      <NxButton onClick={onButtonClick}>Click Me</NxButton>
-    </NxNexusPageHeader>
-  );
-};
-
-export default ComplexPageHeaderExample;
